[zoomer-fade] Fix Radio fade issue related to zoomed text
