+++ packages/perseus/src/widgets/interactive-graphs/graphs/point.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 
 import {actions} from "../reducer/interactive-graph-action";
 import useGraphConfig from "../reducer/use-graph-config";
-import {getCSSZoomFactor} from "../utils";
+import {getCSSZoomFactor} from "../../../util";
 
 import {MovablePoint} from "./components/movable-point";
 import {srFormatNumber} from "./screenreader-text";
